cloud/functions/orchestrator/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.34.*
google-cloud-tasks==2.13.*